requirements.txt
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
pytest==7.4.3
openai==1.3.5
cryptography>=39.0.0